Adjust to the new U2F API
This patch adjusts the various components to the new U2F API, and the
corresponding github.com/tstranex/u2f backwards-incompatible changes.

- The RPC protocol doesn't need change, as we continue to pass opaque
  JSON-serialized requests.
- The proxy's javascript is adjusted, it now does a json unmarshalling
  of the request, and makes the appropriate call to the new API.
- The client library (used mainly by the CLI) got some adjustments for
  the new tstranex/u2f's changes, including API changes.

The configuration format remains unchanged and is backwards compatible,
there's no need to re-register keys.

Unfortunately, only the new CLI will work with the new proxy, and
vice-versa, due to the javascript changes. This could have been avoided
with a significant amount of work but IMO it was not worth the
additional complexity.